# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Book-A-Meal is an app designed to let you order meals from your favourite resturant at your convience. To checkout the app, visit [https://rovilay.github.io/Book-A-Meal/UI](https://rovilay.github.io/Book-A-Meal/UI) .",
    "version" : "1.0.0",
    "title" : "Book-A-Meal",
    "contact" : {
      "email" : "ogooluwaakinola@gmail.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    }
  },
  "host" : "book-me-a-meal.herokuapp.com",
  "basePath" : "/api/v1",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/menus/<menuId>/meals" : {
      "get" : {
        "tags" : [ "Menus" ],
        "summary" : "gets a meals for a menu",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "menuId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "x-example" : "k301c266-b3a3-4a1e-85d7-2a70c45b3042"
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "x-example" : 0
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "x-example" : 2
        } ],
        "responses" : {
          "200" : {
            "description" : "Meals retrieved successfully!",
            "schema" : {
              "$ref" : "#/definitions/Menu_Res_Success"
            }
          },
          "400" : {
            "description" : "An input field is empty or Error occurred while getting all menus!",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "403" : {
            "description" : "User not allowed!",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Menu not found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        },
        "security" : [ {
          "jwtHeader" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Menus" ],
        "summary" : "Adds meals to an existing menu",
        "description" : "This operation can only be performed by an admin(caterer)",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "menuId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "x-example" : "k301c266-b3a3-4a1e-85d7-2a70c45b3042"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Meal_IDs"
          },
          "x-examples" : {
            "application/json" : "{\n    \"meals\": [\"e301c266-b3a3-4a1e-85d7-2a70c45b3042\"]\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Meals added to menu successfully!",
            "schema" : {
              "$ref" : "#/definitions/Success"
            }
          },
          "400" : {
            "description" : "MenuId is incorrect OR <> is empty!",
            "schema" : {
              "$ref" : "#/definitions/Success"
            }
          },
          "403" : {
            "description" : "User not allowed",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Menu not found!",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "405" : {
            "description" : "Can't modify menu anymore",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        },
        "security" : [ {
          "jwtHeader" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Menus" ],
        "summary" : "Removes meals from a menu",
        "description" : "This operation can only be performed by an admin(caterer)",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "menuId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "x-example" : "e301c266-b3a3-4a1e-85d7-2a70c45b3042"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Meal_IDs"
          },
          "x-examples" : {
            "application/json" : "{\n    \"meals\": [\"mealId\"]\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Meals removed from menu successfully!",
            "schema" : {
              "$ref" : "#/definitions/Success"
            }
          },
          "400" : {
            "description" : "MenuId is incorrect OR <> is empty!",
            "schema" : {
              "$ref" : "#/definitions/Success"
            }
          },
          "403" : {
            "description" : "User not allowed!",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Menu not found!",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "405" : {
            "description" : "Can't modify menu anymore",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        },
        "security" : [ {
          "jwtHeader" : [ ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "jwtHeader" : {
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
...
    "Pagination" : {
      "properties" : {
        "limit" : {
          "type" : "integer",
          "format" : "int32"
        },
        "offset" : {
          "type" : "integer",
          "format" : "int32"
        },
        "count" : {
          "type" : "integer",
          "format" : "int32"
        },
        "numOfPages" : {
          "type" : "integer",
          "format" : "int32"
        },
        "curPage" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nextOffset" : {
          "type" : "integer",
          "format" : "int32"
        }
      },
      "example" : {
        "numOfPages" : 5,
        "curPage" : 5,
        "offset" : 6,
        "limit" : 0,
        "count" : 1,
        "nextOffset" : 2
      }
    },
...
    "Success" : {
      "properties" : {
        "success" : {
          "type" : "boolean"
        },
        "message" : {
          "type" : "string"
        }
      },
      "example" : {
        "success" : true,
        "message" : "message"
      }
    },
    "Menu" : {
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "postOn" : {
          "type" : "string"
        },
        "UserId" : {
          "type" : "string"
        },
        "createdAt" : {
          "type" : "string"
        },
        "updatedAt" : {
          "type" : "string"
        },
        "User" : {
          "$ref" : "#/definitions/Menu_Creator"
        },
        "Meals" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Menu_Meals"
          }
        }
      },
      "example" : {
        "createdAt" : "createdAt",
        "User" : {
          "firstName" : "firstName",
          "lastName" : "lastName"
        },
        "UserId" : "UserId",
        "id" : "id",
        "Meals" : [ {
          "image" : "image",
          "price" : 0,
          "MenuMeal" : {
            "id" : 6
          },
          "description" : "description",
          "id" : "id",
          "title" : "title"
        }, {
          "image" : "image",
          "price" : 0,
          "MenuMeal" : {
            "id" : 6
          },
          "description" : "description",
          "id" : "id",
          "title" : "title"
        } ],
        "postOn" : "postOn",
        "updatedAt" : "updatedAt"
      }
    },
...
    "Meal_IDs" : {
      "properties" : {
        "meals" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "description" : "meal ID"
          }
        }
      },
      "example" : {
        "meals" : [ "meals", "meals" ]
      }
    },
    "MenuMeal" : {
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        }
      },
      "description" : "id in menuMeal join table",
      "example" : {
        "id" : 6
      }
    },
    "Menu_Meals" : {
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "price" : {
          "type" : "integer",
          "format" : "int32"
        },
        "image" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "MenuMeal" : {
          "$ref" : "#/definitions/MenuMeal"
        }
      },
      "example" : {
        "image" : "image",
        "price" : 0,
        "MenuMeal" : {
          "id" : 6
        },
        "description" : "description",
        "id" : "id",
        "title" : "title"
      }
    },
...
    "Menu_Creator" : {
      "properties" : {
        "firstName" : {
          "type" : "string"
        },
        "lastName" : {
          "type" : "string"
        }
      },
      "description" : "Caterer who created the menu",
      "example" : {
        "firstName" : "firstName",
        "lastName" : "lastName"
      }
    },
    "Menu_Res_Success" : {
      "properties" : {
        "success" : {
          "type" : "boolean"
        },
        "message" : {
          "type" : "string"
        },
        "pagination" : {
          "$ref" : "#/definitions/Pagination"
        },
        "menu" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Menu"
          }
        }
      },
      "example" : {
        "pagination" : {
          "numOfPages" : 5,
          "curPage" : 5,
          "offset" : 6,
          "limit" : 0,
          "count" : 1,
          "nextOffset" : 2
        },
        "success" : true,
        "message" : "message",
        "menu" : [ {
          "createdAt" : "createdAt",
          "User" : {
            "firstName" : "firstName",
            "lastName" : "lastName"
          },
          "UserId" : "UserId",
          "id" : "id",
          "Meals" : [ {
            "image" : "image",
            "price" : 0,
            "MenuMeal" : {
              "id" : 6
            },
            "description" : "description",
            "id" : "id",
            "title" : "title"
          }, {
            "image" : "image",
            "price" : 0,
            "MenuMeal" : {
              "id" : 6
            },
            "description" : "description",
            "id" : "id",
            "title" : "title"
          } ],
          "postOn" : "postOn",
          "updatedAt" : "updatedAt"
        }, {
          "createdAt" : "createdAt",
          "User" : {
            "firstName" : "firstName",
            "lastName" : "lastName"
          },
          "UserId" : "UserId",
          "id" : "id",
          "Meals" : [ {
            "image" : "image",
            "price" : 0,
            "MenuMeal" : {
              "id" : 6
            },
            "description" : "description",
            "id" : "id",
            "title" : "title"
          }, {
            "image" : "image",
            "price" : 0,
            "MenuMeal" : {
              "id" : 6
            },
            "description" : "description",
            "id" : "id",
            "title" : "title"
          } ],
          "postOn" : "postOn",
          "updatedAt" : "updatedAt"
        } ]
      }
    },
...
    "Error" : {
      "properties" : {
        "success" : {
          "type" : "boolean",
          "default" : false
        },
        "message" : {
          "type" : "string"
        }
      },
      "description" : "Response recieved when operation is unsuccessful"
    }
  },
...
}